requirements.txt
attrs==19.1.0
cached-property==1.5.1
networkx==2.3
numpy==1.16.4
pandas==0.25.3
scikit-learn==0.21.2
scipy==1.3.1
tqdm==4.32.1